Introduce bin/sifon
