+++ bin/sifon
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+// vim: set ts=2 sw=2 sts=2 et ai:
+
+require("../src/cli/cli");
+
