+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "sifon",
   "version": "0.1.0",
+  "private": true,
   "description": "A metaprogrammable altJS",
   "keywords": [
     "compiler",
@@ -9,16 +10,17 @@
   ],
   "author": "kizz",
   "license": "MIT",
-  "main": "src/cui/sifon.js",
+  "bin": "bin/sifon",
+  "main": "src/core/Compiler.js",
   "scripts": {
     "test": "mocha"
   },
   "devDependencies": {
     "modulr": "~0.14.0",
+    "mocha": "~2.2.4",
     "chai": "~1.9.1"
   },
   "dependencies": {
-    "minimist": "~0.1.0",
-    "yargs": "~1.3.1"
+    "commander": "^2.8.1"
   }
 }
